+++ generators/lcg64prime.c
@@ -18,7 +18,6 @@
  */
 #include "smokerand/cinterface.h"
 #include "smokerand/int128defs.h"
-#include <inttypes.h>
 
 PRNG_CMODULE_PROLOG
 
@@ -85,7 +84,8 @@ static int run_self_test(const CallerAPI *intf)
     for (size_t i = 0; i < 100000; i++) {
         u = get_bits_raw(&obj);
     }
-    intf->printf("Result: %" PRIu64 "; reference value: %" PRIu64 "\n", u, u_ref);
+    intf->printf("Result: %llu; reference value: %llu\n",
+        (unsigned long long) u, (unsigned long long) u_ref);
     return u == u_ref;
 }
